+++ src/main/java/com/lets/springproj/domain/LikePostStatus.java
@@ -1,0 +1,5 @@
+package com.lets.springproj.domain;
+
+public enum LikePostStatus {
+    ACTIVE, INACTIVE
+}
