+++ src/main/java/com/lets/springproj/domain/PostStatus.java
@@ -1,0 +1,5 @@
+package com.lets.springproj.domain;
+
+public enum PostStatus {
+    RECRUITING, DONE
+}
